Avoid duplicates and improve mod download logic
The update in the utility function for handling mods prevents duplication of mods and improves the logic relating to downloading mods from Modrinth. Now, mods are only pushed onto the list if they don't already exist there. Changes to the download logic ensures that the correct mod names are returned even when a mod is already downloaded.